requirements.txt
PyYAML==5.1
pytest==5.1.2
mypy
